+++ crates/primitives/starknet/Cargo.toml
@@ -26,7 +26,9 @@ starknet-ff = { workspace = true, default-features = false, features = [
   "alloc",
 ] }
 poseidon_hash = { workspace = true, default-features = false }
-blockifier = { workspace = true, default-features = false }
+blockifier = { workspace = true, default-features = false, features = [
+  "testing",
+] }
 starknet_api = { workspace = true, default-features = false }
 
 # Other third party dependencies
@@ -40,10 +42,12 @@ scale-info = { version = "2.5.0", default-features = false, features = [
 serde = { version = "1.0.159", features = ["derive"], optional = true, default-features = false }
 serde_json = { version = "1.0.95", default-features = false }
 bitvec = { version = "0.17.4", features = ["alloc"], default-features = false }
+hex = { version = "0.4.3", default-features = false }
 
 [dev-dependencies]
 rand = "0.8.5"
 zstd = { version = "0.12.3", default-features = false }
+pretty_assertions = "1.3.0"
 
 [features]
 default = ["std"]
